Speed up sysclk by a factor of two
The previous testbenches inadvertently depended on a phase relationship between
the sysclk and the measured clk signal. Although intuitively I believed that
sysclk should only have to be twice as fast as the measured clk (by Nyquist),
it is more correct to think of the code as needing to resolve half-cycles (in
order to find negative-going edges), and therefore needing a minimum frequency
of 4x the measured clk.

In any case, a 4x multiple of a maximum of 16.7kHz (from the PS/2 interface
definition) is not onerous.